Move all Comps into Component file and change filepaths
