+++ gallery-time/src/Components/FavoriteGalleries/FavoriteGalleries.css
@@ -6,7 +6,7 @@
   display: flex;
   flex-direction: column;
   flex-wrap: wrap;
-  background: url('../img/galleries-background.jpg');
+  background: url('../../img/galleries-background.jpg');
   background-repeat: repeat;
   background-size: cover;
 }
