+++ gallery-time/src/Components/Galleries/Galleries.css
@@ -5,7 +5,7 @@
   display: flex;
   justify-content: space-around;
   flex-wrap: wrap;
-  background: url('../img/galleries-background.jpg');
+  background: url('../../img/galleries-background.jpg');
   background-repeat: repeat;
   background-size: cover;
 }
