+++ gallery-time/src/Components/GalleryDetail/GalleryDetail.css
@@ -5,14 +5,14 @@
   display: flex;
   justify-content: space-around;
   flex-wrap: wrap;
-  background: url('../img/galleries-background.jpg');
+  background: url('../../img/galleries-background.jpg');
   background-repeat: repeat;
   background-size: cover;
 }
 
 .styling-container {
   width: 90vw;
-  background: url('../img/background-card.jpg');
+  background: url('../../img/background-card.jpg');
   background-size: cover;
   box-shadow: 5px 5px 7px 2px rgba(176, 190, 197, .8),
   -2px -2px 5px 1px rgba(176, 190, 197, .4);
