+++ gallery-time/src/Components/LandingPage/LandingPage.css
@@ -1,5 +1,5 @@
 .landing {
-  background: url('../img/background-card.jpg');
+  background: url('../../img/background-card.jpg');
   background-size: cover;
   height: 100vh;
   width: 100vw;
