+++ gallery-time/src/Components/Loading/Loading.js
