+++ gallery-time/src/Components/OffLine/OffLine.css
@@ -1,7 +1,7 @@
 .offline-container {
   height: 100vh;
   width:100vw;
-  background:url('../img/the-scream-1893-by-edvard-munch-116046966.jpeg') repeat
+  background:url('../../img/the-scream-1893-by-edvard-munch-116046966.jpeg') repeat
 }
 
 .offline-p {
